requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0; python_version >= "3.6"
aioschedule==0.5.2
aiosqlite==0.16.1; python_version >= "3.7" and python_version < "4.0"
amqp==5.0.6; python_version >= "3.6"
async-timeout==3.0.1; python_full_version >= "3.5.3" and python_version >= "3.6"
asyncpg==0.23.0; python_full_version >= "3.5.0"
attrs==21.2.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
babel==2.9.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
billiard==3.6.4.0; python_version >= "3.6"
celery==5.1.2; python_version >= "3.6"
certifi==2021.5.30
chardet==4.0.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
click-didyoumean==0.0.3; python_version >= "3.6"
click-plugins==1.1.1; python_version >= "3.6"
click-repl==0.2.0; python_version >= "3.6"
click==7.1.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
idna==3.2; python_version >= "3.6"
iso8601==0.1.16; python_version >= "3.7" and python_version < "4.0"
kombu==5.1.0; python_version >= "3.6"
multidict==5.1.0; python_version >= "3.6"
prompt-toolkit==3.0.19; python_full_version >= "3.6.1" and python_version >= "3.6"
pypika-tortoise==0.1.1; python_version >= "3.7" and python_version < "4.0"
pytz==2021.1; python_version >= "3.7" and python_version < "4.0" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0")
redis==3.5.3; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
six==1.16.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
tortoise-orm==0.17.6; python_version >= "3.7" and python_version < "4.0"
typing-extensions==3.10.0.0; python_version >= "3.7" and python_version < "4.0"
vine==5.0.0; python_version >= "3.6"
wcwidth==0.2.5; python_full_version >= "3.6.1" and python_version >= "3.6"
yarl==1.6.3; python_version >= "3.6"
